refactor: update action metadata and improve documentation
- Renamed action from 'Actions Status Discord' to 'Discord Status' in action.yml and README.md for clarity.
- Enhanced input descriptions in action.yml for better understanding.
- Updated package.json to reflect the new action name and entry point changes.
- Removed outdated CONTRIBUTING.md and issue templates to streamline the repository.
- Added verification and release instructions in AGENTS.md for better developer guidance.
- Updated .gitignore to exclude build artifacts and logs.

These changes aim to improve the overall structure and usability of the action.
