--- .github/CODEOWNERS
@@ -1,1 +1,0 @@
-* @iShark5060
