--- .github/ISSUE_TEMPLATE/bug_report.md
@@ -1,33 +1,0 @@
----
-name: Bug report
-about: Create a report to help us improve
-title: ''
-labels: bug
-assignees: iShark5060
----
-
-**Describe the bug**
-
-<!-- A clear and concise description of what the bug is. -->
-
-**To Reproduce**
-
-<!-- Paste workflow steps, if available: -->
-
-```yaml
-- uses: iShark5060/actions-status-discord@v1
-  with:
-    webhook: ...
-```
-
-**Expected behavior**
-
-<!-- A clear and concise description of what you expected to happen. -->
-
-**Screenshots**
-
-<!-- If applicable, add screenshots to help explain your problem. -->
-
-**Additional context**
-
-<!-- Add any other context about the problem here. -->
