--- .github/ISSUE_TEMPLATE/config.yml
@@ -1,8 +1,0 @@
-blank_issues_enabled: false
-contact_links:
-  - name: Feature requests
-    url: https://github.com/iShark5060/actions-status-discord/discussions/new?category=ideas-feature-requests
-    about: Please use GitHub Discussions for feature requests!
-  - name: Q&A
-    url: https://github.com/iShark5060/actions-status-discord/discussions/new?category=q-a
-    about: Please use GitHub Discussions for Q&A!
