+++ .github/workflows/ci.yml
@@ -46,4 +46,4 @@ jobs:
         run: pnpm run build
 
       - name: Verify action bundle
-        run: node --check lib/index.js
+        run: node --check dist/index.js
