+++ .gitignore
@@ -1,4 +1,8 @@
 .DS_Store
 /.idea
 /node_modules
+/dist
 /lib
+*.log
+coverage/
+.nyc_output/
