--- CONTRIBUTING.md
@@ -1,14 +1,0 @@
-# Developer Docs
-
-## Release
-
-1. Make **pre-release** from `main` branch
-2. Check e2e action result
-3. Release again!
-
-## Update dependencies
-
-```
-$ pnpm run deps
-$ pnpm install
-```
