+++ action.yml
@@ -1,76 +1,76 @@
-name: 'Actions Status Discord'
-description: 'Post Github Actions CI status to Discord'
-author: 'Sarisia'
+name: Discord Status
+description: Post GitHub Actions CI status to Discord
+author: iShark5060
 
 inputs:
   webhook:
-    description: 'Discord webhook endpoint. If not set, env.DISCORD_WEBHOOK will be used.'
+    description: Discord webhook endpoint. If not set, env.DISCORD_WEBHOOK will be used.
     required: false
   status:
-    description: 'Job status. Should be bound to job.status or workflow conclusion. Default to success.'
+    description: Job status. Should be bound to job.status or workflow conclusion. Default to success.
     required: false
     default: ${{ job.status }}
   job:
-    description: 'Deprecated. Job name included in message title. Same as title input.'
+    description: Deprecated. Job name included in message title. Same as title input.
     required: false
   content:
-    description: 'Content. Shown as an message outside of the embed. See [Mention to user/role](#mention-to-user-role)'
+    description: Content shown as a message outside of the embed.
     required: false
   title:
-    description: 'String included in embed title. Overrides job input.'
+    description: String included in embed title. Overrides job input.
     required: false
     default: ${{ github.workflow }}
   description:
-    description: 'Description included in message'
+    description: Description included in message
     required: false
   image:
-    description: 'Image attached to the message'
+    description: Image attached to the message
     required: false
   color:
-    description: 'Overrides Discord embed color'
+    description: Overrides Discord embed color
     required: false
   url:
-    description: 'URL to jump when the title is clicked'
+    description: URL to jump when the title is clicked
     required: false
   username:
-    description: 'Overrides Discord webhook username'
+    description: Overrides Discord webhook username
     required: false
   avatar_url:
-    description: 'Overrides Discord webhook avatar url'
+    description: Overrides Discord webhook avatar url
     required: false
   nofail:
-    description: "This action won't make workflow failed by default."
+    description: This action won't make workflow failed by default.
     required: false
     default: 'true'
   nocontext:
-    description: 'Suppress GitHub context fields'
+    description: Suppress GitHub context fields
     required: false
     default: 'false'
   noprefix:
-    description: 'Avoid appending job status to title'
+    description: Avoid appending job status to title
     required: false
     default: 'false'
   nodetail:
-    description: 'Suppress detailed embed fields'
+    description: Suppress detailed embed fields
     required: false
     default: 'false'
   notimestamp:
-    description: 'Avoid appending timestamp'
+    description: Avoid appending timestamp
     required: false
     default: 'false'
   ack_no_webhook:
-    description: 'Suppress error which raised when webhook is not set'
+    description: Suppress error which raised when webhook is not set
     required: false
     default: 'false'
 
 outputs:
   payload:
-    description: 'Discord webhook payload'
+    description: Discord webhook payload
 
 runs:
-  using: 'node24'
-  main: 'lib/index.js'
+  using: node24
+  main: dist/index.js
 
 branding:
-  color: 'blue'
-  icon: 'send'
+  color: blue
+  icon: send
