--- lib-test/index.js
@@ -1,1 +1,0 @@
-(()=>{"use strict";var e={47:(e,t)=>{Object.defineProperty(t,"__esModule",{value:true});t.MAX_WEBHOOK_CONTENT_LENGTH=t.MAX_EMBED_FIELD_VALUE_LENGTH=t.MAX_EMBED_FIELD_NAME_LENGTH=t.MAX_EMBED_DESCRIPTION_LENGTH=t.MAX_EMBED_TITLE_LENGTH=void 0;t.MAX_EMBED_TITLE_LENGTH=256;t.MAX_EMBED_DESCRIPTION_LENGTH=4096;t.MAX_EMBED_FIELD_NAME_LENGTH=256;t.MAX_EMBED_FIELD_VALUE_LENGTH=1024;t.MAX_WEBHOOK_CONTENT_LENGTH=2e3},905:(e,t,n)=>{Object.defineProperty(t,"__esModule",{value:true});t.getContext=getContext;const r=n(896);function readPayload(){const e=process.env.GITHUB_EVENT_PATH;if(e&&(0,r.existsSync)(e)){return JSON.parse((0,r.readFileSync)(e,{encoding:"utf8"}))}return{}}function readRepo(e){if(process.env.GITHUB_REPOSITORY){const[e,t]=process.env.GITHUB_REPOSITORY.split("/");return{owner:e,repo:t}}if(e.repository){return{owner:e.repository.owner.login,repo:e.repository.name}}return{owner:"",repo:""}}function getContext(){const e=readPayload();return{payload:e,eventName:process.env.GITHUB_EVENT_NAME||"",ref:process.env.GITHUB_REF||"",workflow:process.env.GITHUB_WORKFLOW||"",actor:process.env.GITHUB_ACTOR||"",serverUrl:process.env.GITHUB_SERVER_URL||"https://github.com",runId:parseInt(process.env.GITHUB_RUN_ID||"0",10),repo:readRepo(e)}}},231:(e,t,n)=>{Object.defineProperty(t,"__esModule",{value:true});t.formatEvent=formatEvent;const r=n(275);const o={push:pushFormatter,pull_request:pullRequestFormatter,release:releaseFormatter};function formatEvent(e,t){(0,r.logDebug)(JSON.stringify(t,null,2));let n="No further information";if(e in o){try{return o[e](t)||n}catch(t){(0,r.logDebug)(`Failed to generate eventDetail for ${e}: ${t}\n${t.stack}`)}}return n}function pushFormatter(e){return`[\`${e.head_commit.id.substring(0,7)}\`](${e.head_commit.url}) ${e.head_commit.message}`}function pullRequestFormatter(e){return`[\`#${e.pull_request.number}\`](${e.pull_request.html_url}) ${e.pull_request.title}`}function releaseFormatter(e){const{name:t,body:n}=e.release;const r=t?`**${t}**`:"";return`${r}${r&&n?"\n":""}${n||""}`}},750:function(e,t,n){var r=this&&this.__createBinding||(Object.create?function(e,t,n,r){if(r===undefined)r=n;var o=Object.getOwnPropertyDescriptor(t,n);if(!o||("get"in o?!t.__esModule:o.writable||o.configurable)){o={enumerable:true,get:function(){return t[n]}}}Object.defineProperty(e,r,o)}:function(e,t,n,r){if(r===undefined)r=n;e[r]=t[n]});var o=this&&this.__setModuleDefault||(Object.create?function(e,t){Object.defineProperty(e,"default",{enumerable:true,value:t})}:function(e,t){e["default"]=t});var i=this&&this.__importStar||function(){var ownKeys=function(e){ownKeys=Object.getOwnPropertyNames||function(e){var t=[];for(var n in e)if(Object.prototype.hasOwnProperty.call(e,n))t[t.length]=n;return t};return ownKeys(e)};return function(e){if(e&&e.__esModule)return e;var t={};if(e!=null)for(var n=ownKeys(e),i=0;i<n.length;i++)if(n[i]!=="default")r(t,e,n[i]);o(t,e);return t}}();Object.defineProperty(t,"__esModule",{value:true});t.statusOpts=void 0;t.getInputs=getInputs;const u=i(n(Object(function webpackMissingModule(){var e=new Error("Cannot find module '@actions/core'");e.code="MODULE_NOT_FOUND";throw e}())));const s=n(275);t.statusOpts={success:{status:"Success",color:2664261},failure:{status:"Failure",color:13313073},skipped:{status:"Skipped",color:9804188},cancelled:{status:"Cancelled",color:14396169}};function getInputs(){const e=u.getInput("webhook").trim()||process.env.DISCORD_WEBHOOK||"";const n=e.split("\n").filter(e=>e||false);n.forEach((e,t)=>{u.setSecret(e);if(e.endsWith("/github")){(0,s.logWarning)(`webhook ${t+1}/${n.length} has \`/github\` suffix! This may cause errors.`)}});const r=(0,s.stob)(u.getInput("nodetail"));const o=r||(0,s.stob)(u.getInput("nocontext"));const i=r||(0,s.stob)(u.getInput("noprefix"));const a=parseInt(u.getInput("color"));const c={webhooks:n,status:u.getInput("status").trim().toLowerCase(),content:u.getInput("content").trim(),description:u.getInput("description").trim(),title:(u.getInput("job")||u.getInput("title")).trim(),image:u.getInput("image").trim(),color:isNaN(a)?undefined:a,url:u.getInput("url").trim(),username:u.getInput("username").trim(),avatar_url:u.getInput("avatar_url").trim(),nocontext:o,noprefix:i,notimestamp:(0,s.stob)(u.getInput("notimestamp")),ack_no_webhook:(0,s.stob)(u.getInput("ack_no_webhook"))};if(!c.webhooks.length&&!c.ack_no_webhook){throw new Error("No webhook is given. If this is intended, you can suppress this error by setting `ack_no_webhook` to `true`.")}if(!(c.status in t.statusOpts)){throw new Error(`invalid status value: ${c.status}`)}return c}},275:function(e,t,n){var r=this&&this.__createBinding||(Object.create?function(e,t,n,r){if(r===undefined)r=n;var o=Object.getOwnPropertyDescriptor(t,n);if(!o||("get"in o?!t.__esModule:o.writable||o.configurable)){o={enumerable:true,get:function(){return t[n]}}}Object.defineProperty(e,r,o)}:function(e,t,n,r){if(r===undefined)r=n;e[r]=t[n]});var o=this&&this.__setModuleDefault||(Object.create?function(e,t){Object.defineProperty(e,"default",{enumerable:true,value:t})}:function(e,t){e["default"]=t});var i=this&&this.__importStar||function(){var ownKeys=function(e){ownKeys=Object.getOwnPropertyNames||function(e){var t=[];for(var n in e)if(Object.prototype.hasOwnProperty.call(e,n))t[t.length]=n;return t};return ownKeys(e)};return function(e){if(e&&e.__esModule)return e;var t={};if(e!=null)for(var n=ownKeys(e),i=0;i<n.length;i++)if(n[i]!=="default")r(t,e,n[i]);o(t,e);return t}}();Object.defineProperty(t,"__esModule",{value:true});t.logError=logError;t.logDebug=logDebug;t.logInfo=logInfo;t.logWarning=logWarning;t.stob=stob;const u=i(n(Object(function webpackMissingModule(){var e=new Error("Cannot find module '@actions/core'");e.code="MODULE_NOT_FOUND";throw e}())));const s=u.getInput("nofail").trim().toLowerCase()==="true";function logError(e){s?u.error(e):u.setFailed(e)}function logDebug(e){u.debug(e)}function logInfo(e){u.info(e)}function logWarning(e){u.warning(e)}function stob(e){return e.trim().toLowerCase()==="true"}},998:function(e,t,n){var r=this&&this.__createBinding||(Object.create?function(e,t,n,r){if(r===undefined)r=n;var o=Object.getOwnPropertyDescriptor(t,n);if(!o||("get"in o?!t.__esModule:o.writable||o.configurable)){o={enumerable:true,get:function(){return t[n]}}}Object.defineProperty(e,r,o)}:function(e,t,n,r){if(r===undefined)r=n;e[r]=t[n]});var o=this&&this.__setModuleDefault||(Object.create?function(e,t){Object.defineProperty(e,"default",{enumerable:true,value:t})}:function(e,t){e["default"]=t});var i=this&&this.__importStar||function(){var ownKeys=function(e){ownKeys=Object.getOwnPropertyNames||function(e){var t=[];for(var n in e)if(Object.prototype.hasOwnProperty.call(e,n))t[t.length]=n;return t};return ownKeys(e)};return function(e){if(e&&e.__esModule)return e;var t={};if(e!=null)for(var n=ownKeys(e),i=0;i<n.length;i++)if(n[i]!=="default")r(t,e,n[i]);o(t,e);return t}}();Object.defineProperty(t,"__esModule",{value:true});t.truncStr=truncStr;t.fitEmbed=fitEmbed;t.fitContent=fitContent;const u=i(n(47));const s=n(275);function truncStr(e,t){return e.slice(0,t-3)+"..."}function fitEmbed(e){if(e.title){const t=e.title.length;if(t>u.MAX_EMBED_TITLE_LENGTH){(0,s.logWarning)(`embed title must be shorter than ${u.MAX_EMBED_TITLE_LENGTH}, got ${t}\n    ${e.title}`);e.title=truncStr(e.title,u.MAX_EMBED_TITLE_LENGTH)}}if(e.description){const t=e.description.length;if(t>u.MAX_EMBED_DESCRIPTION_LENGTH){(0,s.logWarning)(`embed description must be shorter than ${u.MAX_EMBED_DESCRIPTION_LENGTH}, got ${t}\n    ${e.description}`);e.description=truncStr(e.description,u.MAX_EMBED_DESCRIPTION_LENGTH)}}if(e.fields){for(const t of e.fields){const e=t.name.length;const n=t.value.length;if(e>u.MAX_EMBED_FIELD_NAME_LENGTH){(0,s.logWarning)(`embed field name must be shorter than ${u.MAX_EMBED_FIELD_NAME_LENGTH}, got ${e}\n    ${t.name}`);t.name=truncStr(t.name,u.MAX_EMBED_FIELD_NAME_LENGTH)}if(n>u.MAX_EMBED_FIELD_VALUE_LENGTH){(0,s.logWarning)(`embed field value must be shorter than ${u.MAX_EMBED_FIELD_VALUE_LENGTH}, got ${n}\n    ${t.value}`);t.value=truncStr(t.value,u.MAX_EMBED_FIELD_VALUE_LENGTH)}}}return e}function fitContent(e){const t=e.length;if(t>u.MAX_WEBHOOK_CONTENT_LENGTH){(0,s.logWarning)(`content field must be shorter than ${u.MAX_WEBHOOK_CONTENT_LENGTH}, got ${t}\n    ${e}`);e=truncStr(e,u.MAX_WEBHOOK_CONTENT_LENGTH)}return e}},896:e=>{e.exports=require("fs")}};var t={};function __nccwpck_require__(n){var r=t[n];if(r!==undefined){return r.exports}var o=t[n]={exports:{}};var i=true;try{e[n].call(o.exports,o,o.exports,__nccwpck_require__);i=false}finally{if(i)delete t[n]}return o.exports}if(typeof __nccwpck_require__!=="undefined")__nccwpck_require__.ab=__dirname+"/";var n={};(()=>{var e=n;Object.defineProperty(e,"__esModule",{value:true});e.getPayload=getPayload;const t=__nccwpck_require__(Object(function webpackMissingModule(){var e=new Error("Cannot find module '@actions/core'");e.code="MODULE_NOT_FOUND";throw e}()));const r=__nccwpck_require__(905);const o=__nccwpck_require__(231);const i=__nccwpck_require__(750);const u=__nccwpck_require__(275);const s=__nccwpck_require__(998);async function run(){try{(0,u.logInfo)("Getting inputs...");const e=(0,i.getInputs)();(0,u.logInfo)("Generating payload...");const n=getPayload(e);const r=JSON.stringify(n,null,2);(0,t.startGroup)("Dump payload (You can access the payload as `${{ steps.<step_id>.outputs.payload }}` in latter steps)");(0,u.logInfo)(r);(0,t.endGroup)();(0,u.logInfo)(`Triggering ${e.webhooks.length} webhook${e.webhooks.length>1?"s":""}...`);const o=await Promise.allSettled(e.webhooks.map(e=>wrapWebhook(e.trim(),n)));const s=o.filter(e=>e.status==="rejected");if(s.length>0){for(const e of s){const t=e.reason instanceof Error?e.reason.message:String(e.reason);(0,u.logError)(t)}return}(0,t.setOutput)("payload",r)}catch(e){const t=e instanceof Error?e.message:String(e);(0,u.logError)(`Unexpected failure: ${t}`)}}async function wrapWebhook(e,t){const n=await fetch(e,{method:"POST",headers:{"Content-Type":"application/json"},body:JSON.stringify(t)});if(!n.ok){const e=await n.text();throw new Error(`Webhook response: ${n.status}: ${e}`)}}function getPayload(e){const t=(0,r.getContext)();const{owner:n,repo:a}=t.repo;const{eventName:c,ref:l,workflow:_,actor:f,payload:p,serverUrl:d,runId:g}=t;const E=`${d}/${n}/${a}`;const b=`${E}/actions/runs/${g}`;(0,u.logDebug)(JSON.stringify(p));const m=`Event - ${c}`;const h=(0,o.formatEvent)(c,p);let O={color:e.color===undefined?i.statusOpts[e.status].color:e.color};if(!e.notimestamp){O.timestamp=(new Date).toISOString()}if(e.title){O.title=e.title}if(e.url){O.url=e.url}if(e.image){O.image={url:e.image}}if(!e.noprefix){O.title=i.statusOpts[e.status].status+(O.title?`: ${O.title}`:"")}if(e.description){O.description=e.description}if(!e.nocontext){O.fields=[{name:"Repository",value:`[${n}/${a}](${E})`,inline:true},{name:"Ref",value:l,inline:true},{name:m,value:h,inline:false},{name:"Triggered by",value:f,inline:true},{name:"Workflow",value:`[${_}](${b})`,inline:true}]}let v={embeds:[(0,s.fitEmbed)(O)]};(0,u.logDebug)(`embed: ${JSON.stringify(O)}`);if(e.username){v.username=e.username}if(e.avatar_url){v.avatar_url=e.avatar_url}if(e.content){v.content=(0,s.fitContent)(e.content)}return v}if(!process.env.VITEST){run()}})();module.exports=n})();
